Drop the work-around for the broken link https://git-scm.com/docs上查看
In c902ac87c (ci: check for broken git-scm.com links, 2025-02-27), I
added a work-around for a link that had been broken in advertently, in a
repository different from git-scm.com. I contributed a fix for that with
https://github.com/jnavila/git-manpages-l10n/pull/131, but did not want
to have git-scm.com's link check to keep failing until that fix was
merged.

Well, the work-around is no longer necessary because my fix has been
merged in the meantime.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>