html-to-pdf: work around minified `application.{css,js}`
When Hugo minifies resources, it attaches an `integrity` attribute. This
is usually very good! However, when we're trying to generate a `.pdf`
from a HTML file that contains hrefs that contain absolute paths, as we
do when deploying to GitHub Pages, we have to work around that by
re-routing the files. Combined with an `integrity` attribute, this will
fail, and there is no way around it other than editing out that
attribute.

So let's do just that.

Hacky, but works, and we can proceed activities that are substantially
more fun than battling well-intentioned-but-sometimes-really-unhelpful
security measures without escape hatches (there is no way to tell
Chromium that yes, this is intentional, but we need to ignore those
integrity checks in this context, thank you for your cooperation).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>